Fixed the inf type problem
Used Oleg's implementation, which is much nicer in a lot of ways (most
notably you can simply use global prompts, which allows for such a nice
way of adding complex control structures). The implementation itself is
still a mistery, but I'll crack it eventually. The idea is basically
almost the same as Peyton's. Though it is based on resumable exceptions,
and not on the continuation monad itself, which might be super
interesting, and also super important is the fact that I have a CPS
transform of this implementation. I'm sure this will be handy.